Moved smartdashboard related code to new class
